requirements.txt
Flask
Flask-Cors
requests
python-dotenv
gunicorn  # Necessary for running Flask in a production environment
